// src/app/shared/company.interface.ts
export interface CompanyI {
    uid?: string;
    sector: string;
    categoria: string;
    subcategoria: string;
    email: string;
    name: string;
    representative: string;
    numNit: string;
    imgNit: string;
    address: string;
    city: string;
    country: string;
    phone: string;
    password?: string;
    roll: string;
    fileRef?: string;
}