// src/app/shared/user.interface.ts
export interface UserI {
    uid?: string;
    sector: string;
    categoria: string;
    subcategoria: string;
    email: string;
    name: string;
    document: string;
    numDocument: string;
    imgDocument: string;
    address: string;
    city: string;
    country: string;
    phone: string;
    password?: string;
    roll: string;
    fileRef?: string;
    relationship: string;
}